The eternal question of how we make dev and prod play nice.
The previous solution (using --assume-unchanged) was a godawful hack,
and when I wrote it I had no idea that it'd make switching branches a
living hell. Now that we're branching out to running Orient code on three,
possibly four machines, with Git newbies trying to navigate the waters, it's time
for something a little less janky. I think this is a reasonably secure solution
to database and server root settings, but it might not play nice with all Apache
versions. We'll find out!